resolved pr comments
