+++ likeminds-feed-android-core/src/main/java/com/likeminds/feed/android/core/activityfeed/view/LMFeedActivityFeedActivity.kt
@@ -1,18 +1,11 @@
 package com.likeminds.feed.android.core.activityfeed.view
 
-import android.annotation.SuppressLint
 import android.content.Context
 import android.content.Intent
-import android.os.Build
 import android.os.Bundle
-import android.view.WindowInsets
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat
 import androidx.core.view.ViewCompat
-import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
-import androidx.core.view.WindowInsetsControllerCompat
-import com.likeminds.feed.android.core.R
 import com.likeminds.feed.android.core.databinding.LmFeedActivityActivityFeedBinding
 
 class LMFeedActivityFeedActivity : AppCompatActivity() {
@@ -38,7 +31,6 @@ class LMFeedActivityFeedActivity : AppCompatActivity() {
         binding = LmFeedActivityActivityFeedBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        setStatusBarColor()
         ViewCompat.setOnApplyWindowInsetsListener(binding.clActivityFeed) { view, windowInsets ->
             val innerPadding = windowInsets.getInsets(
                 // Notice we're using systemBars, not statusBar
@@ -58,23 +50,6 @@ class LMFeedActivityFeedActivity : AppCompatActivity() {
 
         //inflates activity feed fragment
         inflateActivityFeedFragment()
-    }
-
-    @SuppressLint("InlinedApi")
-    @Suppress("Deprecation")
-    private fun setStatusBarColor() {
-        if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
-            val windowInsetsController = WindowCompat.getInsetsController(window, window.decorView)
-            windowInsetsController.isAppearanceLightStatusBars = true
-            window.statusBarColor = ContextCompat.getColor(this, R.color.lm_feed_white)
-        } else {
-            val insetsController = WindowInsetsControllerCompat(window, window.decorView)
-            insetsController.isAppearanceLightStatusBars = true
-
-            window.decorView.setBackgroundColor(ContextCompat.getColor(this, R.color.lm_feed_white))
-            window.insetsController?.hide(WindowInsets.Type.statusBars())
-            window.insetsController?.show(WindowInsets.Type.statusBars())
-        }
     }
 
     private fun inflateActivityFeedFragment() {
