+++ likeminds-feed-android-core/src/main/java/com/likeminds/feed/android/core/poll/create/view/LMFeedCreatePollFragment.kt
@@ -3,9 +3,13 @@ package com.likeminds.feed.android.core.poll.create.view
 import android.annotation.SuppressLint
 import android.app.Activity
 import android.content.Intent
+import android.os.Build
 import android.os.Bundle
 import android.util.Log
 import android.view.*
+import androidx.core.content.ContextCompat
+import androidx.core.view.WindowCompat
+import androidx.core.view.WindowInsetsControllerCompat
 import androidx.core.widget.addTextChangedListener
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
@@ -120,10 +124,34 @@ open class LMFeedCreatePollFragment : Fragment(), LMFeedCreatePollOptionAdapterL
     protected open fun customizeCreatePollHeader(headerView: LMFeedHeaderView) {
         headerView.apply {
             setStyle(LMFeedStyleTransformer.createPollFragmentViewStyle.headerViewStyle)
+            setStatusBarColor(LMFeedStyleTransformer.createPollFragmentViewStyle.headerViewStyle.backgroundColor)
 
             setTitleText(getString(R.string.lm_feed_new_poll))
             setSubmitText(getString(R.string.lm_feed_done))
             setSubmitButtonEnabled(false)
+        }
+    }
+
+    @SuppressLint("InlinedApi")
+    @Suppress("Deprecation")
+    private fun setStatusBarColor(backgroundColor: Int) {
+        val window = requireActivity().window
+        if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
+            val windowInsetsController = WindowCompat.getInsetsController(window, window.decorView)
+            windowInsetsController.isAppearanceLightStatusBars = true
+            window.statusBarColor = ContextCompat.getColor(requireContext(), backgroundColor)
+        } else {
+            val insetsController = WindowInsetsControllerCompat(window, window.decorView)
+            insetsController.isAppearanceLightStatusBars = true
+
+            window.decorView.setBackgroundColor(
+                ContextCompat.getColor(
+                    requireContext(),
+                    backgroundColor
+                )
+            )
+            window.insetsController?.hide(WindowInsets.Type.statusBars())
+            window.insetsController?.show(WindowInsets.Type.statusBars())
         }
     }
