+++ likeminds-feed-android-core/src/main/java/com/likeminds/feed/android/core/post/create/view/LMFeedCreatePostActivity.kt
@@ -1,18 +1,12 @@
 package com.likeminds.feed.android.core.post.create.view
 
-import android.annotation.SuppressLint
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
-import android.os.Build
 import android.os.Bundle
-import android.view.WindowInsets
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat
 import androidx.core.view.ViewCompat
-import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
-import androidx.core.view.WindowInsetsControllerCompat
 import com.likeminds.feed.android.core.R
 import com.likeminds.feed.android.core.databinding.LmFeedActivityCreatePostBinding
 import com.likeminds.feed.android.core.post.create.model.LMFeedCreatePostExtras
@@ -58,7 +52,6 @@ class LMFeedCreatePostActivity : AppCompatActivity() {
         binding = LmFeedActivityCreatePostBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        setStatusBarColor()
         ViewCompat.setOnApplyWindowInsetsListener(binding.clCreatePost) { view, windowInsets ->
             val innerPadding = windowInsets.getInsets(
                 // Notice we're using systemBars, not statusBar
@@ -83,23 +76,6 @@ class LMFeedCreatePostActivity : AppCompatActivity() {
 
         //inflate create post fragment
         inflateCreatePostFragment()
-    }
-
-    @SuppressLint("InlinedApi")
-    @Suppress("Deprecation")
-    private fun setStatusBarColor() {
-        if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
-            val windowInsetsController = WindowCompat.getInsetsController(window, window.decorView)
-            windowInsetsController.isAppearanceLightStatusBars = true
-            window.statusBarColor = ContextCompat.getColor(this, R.color.lm_feed_white)
-        } else {
-            val insetsController = WindowInsetsControllerCompat(window, window.decorView)
-            insetsController.isAppearanceLightStatusBars = true
-
-            window.decorView.setBackgroundColor(ContextCompat.getColor(this, R.color.lm_feed_white))
-            window.insetsController?.hide(WindowInsets.Type.statusBars())
-            window.insetsController?.show(WindowInsets.Type.statusBars())
-        }
     }
 
     //check for extras and handle the flow
