+++ likeminds-feed-android-core/src/main/java/com/likeminds/feed/android/core/search/view/LMFeedSearchActivity.kt
@@ -1,17 +1,11 @@
 package com.likeminds.feed.android.core.search.view
 
-import android.annotation.SuppressLint
 import android.content.Context
 import android.content.Intent
-import android.os.Build
 import android.os.Bundle
-import android.view.WindowInsets
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat
 import androidx.core.view.ViewCompat
-import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
-import androidx.core.view.WindowInsetsControllerCompat
 import com.likeminds.feed.android.core.R
 import com.likeminds.feed.android.core.databinding.LmFeedSearchActivityBinding
 import com.likeminds.feed.android.core.post.create.view.LMFeedCreatePostActivity
@@ -53,7 +47,6 @@ open class LMFeedSearchActivity : AppCompatActivity() {
         binding = LmFeedSearchActivityBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        setStatusBarColor()
         ViewCompat.setOnApplyWindowInsetsListener(binding.clSearch) { view, windowInsets ->
             val innerPadding = windowInsets.getInsets(
                 // Notice we're using systemBars, not statusBar
@@ -78,23 +71,6 @@ open class LMFeedSearchActivity : AppCompatActivity() {
 
         //inflates search feed fragment
         inflateSearchFeedFragment()
-    }
-
-    @SuppressLint("InlinedApi")
-    @Suppress("Deprecation")
-    private fun setStatusBarColor() {
-        if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
-            val windowInsetsController = WindowCompat.getInsetsController(window, window.decorView)
-            windowInsetsController.isAppearanceLightStatusBars = true
-            window.statusBarColor = ContextCompat.getColor(this, R.color.lm_feed_white)
-        } else {
-            val insetsController = WindowInsetsControllerCompat(window, window.decorView)
-            insetsController.isAppearanceLightStatusBars = true
-
-            window.decorView.setBackgroundColor(ContextCompat.getColor(this, R.color.lm_feed_white))
-            window.insetsController?.hide(WindowInsets.Type.statusBars())
-            window.insetsController?.show(WindowInsets.Type.statusBars())
-        }
     }
 
     //check for extras and handle the flow
