+++ lm-feed-root-dependencies.gradle
@@ -4,9 +4,9 @@ feed_versions.compile_sdk = 35
 feed_versions.min_sdk = 21
 feed_versions.target_sdk = 35
 feed_versions.sdk_version_major = '1'
-feed_versions.sdk_version_minor = '105'
+feed_versions.sdk_version_minor = '9'
 feed_versions.sdk_version_patch = '0'
-feed_versions.sdk_version_code = 30
+feed_versions.sdk_version_code = 31
 feed_versions.core = "1.8.0"
 feed_versions.appcompat = "1.6.1"
 feed_versions.material_design = "1.6.1"
